Use UNSAFE.getLong() to speed up BitSetMethods#anySet()

+++ unsafe/src/main/java/org/apache/spark/unsafe/bitset/BitSetMethods.java
@@ -72,7 +72,7 @@ public final class BitSetMethods {
    * Returns {@code true} if any bit is set.
    */
   public static boolean anySet(Object baseObject, long baseOffset, long bitSetWidthInBytes) {
-    long widthInLong = bitSetWidthInBytes / SIZE_OF_LONG;
+    int widthInLong = (int)(bitSetWidthInBytes / SIZE_OF_LONG);
     for (int i = 0; i <= widthInLong; i++) {
       if (PlatformDependent.UNSAFE.getLong(baseObject, baseOffset + i) != 0) {
         return true;
